Improves benchmarking timeout
Timeout check now verifies whether any Docker
container is running on the VM. Before that it was
checked only if the VM itself is still running.
However, in case of a failure the VM will not
automatically shut down to enable debugging of the
error so the whole benchmarking job will be
blocked.

Included an error message that is sent to Slack
in case that the timeout is reached.